requirements.txt
# FastAPI and server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# LLM Providers
openai>=1.3.0
anthropic>=0.7.0

# HTTP client (for async)
httpx>=0.25.0

# CORS
python-multipart>=0.0.6

# Vector Store
chromadb>=0.4.0  # Optional fallback
pgvector>=0.3.0  # PostgreSQL vector extension

# Document Processing
PyPDF2>=3.0.0

# NumPy for embeddings
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0  # PostgreSQL driver (optional)
